requirements.txt
numpy
pandas
keras
tensorflow
dask[complete]
psutil
PyYAML
sklearn
gensim
cython
ujson